chore: tighten lints, CI matrix, repo hygiene, extract query job state
- Narrow the crate-wide clippy allow list from 25 lints to 3 structural ones
  and fix what they were hiding (clamp, struct init, late init, docs, ...).
- CI: clippy -D warnings for all targets with and without `collab` plus tests
  on Linux and macOS (Windows non-blocking), dependency caching, cancel
  superseded runs, and an informational rustfmt check; add rustfmt.toml.
- Drop cargo-deb from runtime dependencies (a packaging CLI, never used by
  the code) and pin mssql-client/mssql-driver-pool to 0.20.x as intended.
- Stop tracking .env (add .env.example), stop ignoring Cargo.lock, and move
  old planning/audit documents from the repo root to docs/archive.
- Group the nine query-job fields of Tabular into QueryJobsState with a
  single allocate_id() in place of eleven copies of the id bookkeeping.
- Honour RUST_LOG only when it holds a valid level spec.

+++ src/syntax_ts.rs
@@ -424,7 +424,7 @@ mod ts {
                 }
 
                 for i in (0..node.child_count()).rev() {
-                    if let Some(child) = node.child(i as u32) {
+                    if let Some(child) = node.child(i) {
                         stack.push(child);
                     }
                 }
@@ -536,7 +536,7 @@ mod ts {
                     });
                 }
                 for i in (0..node.child_count()).rev() {
-                    if let Some(child) = node.child(i as u32) {
+                    if let Some(child) = node.child(i) {
                         stack.push(child);
                     }
                 }
@@ -734,7 +734,7 @@ mod ts {
                 }
 
                 for i in (0..node.child_count()).rev() {
-                    if let Some(child) = node.child(i as u32) {
+                    if let Some(child) = node.child(i) {
                         stack.push(child);
                     }
                 }
@@ -798,7 +798,7 @@ mod ts {
                     });
                 }
                 for i in (0..node.child_count()).rev() {
-                    if let Some(child) = node.child(i as u32) {
+                    if let Some(child) = node.child(i) {
                         stack.push(child);
                     }
                 }
@@ -957,7 +957,7 @@ mod ts {
                 }
 
                 for i in (0..node.child_count()).rev() {
-                    if let Some(child) = node.child(i as u32) {
+                    if let Some(child) = node.child(i) {
                         stack.push(child);
                     }
                 }
@@ -1044,7 +1044,7 @@ mod ts {
                     });
                 }
                 for i in (0..node.child_count()).rev() {
-                    if let Some(child) = node.child(i as u32) {
+                    if let Some(child) = node.child(i) {
                         stack.push(child);
                     }
                 }
